+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `displayAsCurrency` to form element schema
+- `storageType` to form element schema
 
 ## [0.9.9] - 2021-04-27